# Speed up clip_gradient with GPU training