Roughly 25% of GPU time when training a transformer tagger with mixed-precision was spent on computing Frobenius norms for gradient clipping. The reason turns out to be that CuPy only uses a single block for scalar reductions (unless another backend such as CUB is used), resulting in a slow sum reduction when computing a norm.

This change replaces `cupy.linalg.norm` by `cupy.cublas.nrm2`, the optimized cuBLAS implementation for computing l2 norms. This reduces training time about by ~24%.

Top-5 cost before this change:

```
             Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:   28.49%  23.0423s    185554  124.18us  1.0550us  19.842ms  cupy_sum
                    5.68%  4.59333s     42324  108.53us  29.151us  2.6026ms  turing_fp16_s1688gemm_fp16_128x128_ldg8_f2f_tn
                    4.69%  3.78947s    202426  18.720us  1.6320us  1.1343ms  _ZN2at6native27unrolled_elementwise_kernelIZZZNS0_21copy_device_to_deviceERNS_14TensorIteratorEbENKUlvE0_clEvENKUlvE18_clEvEUlN3c104HalfEE_NS_6detail5ArrayIPcLi2EEE23TrivialOffsetCalculatorILi1EjESE_NS0_6memory12LoadWithCastILi1EEENSF_13StoreWithCastEEEviT_T0_T1_T2_T3_T4_
                    4.66%  3.77189s     31488  119.79us  4.5760us  354.82us  turing_fp16_s1688gemm_fp16_128x128_ldg8_f2f_stages_32x1_nt
                    4.51%  3.64793s     79206  46.056us  1.6000us  2.8409ms  adam
```

After:

```
             Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:    7.95%  4.59139s     42324  108.48us  29.184us  2.5801ms  turing_fp16_s1688gemm_fp16_128x128_ldg8_f2f_tn
                    6.56%  3.78955s    202426  18.720us  1.6320us  1.1526ms  _ZN2at6native27unrolled_elementwise_kernelIZZZNS0_21copy_device_to_deviceERNS_14TensorIteratorEbENKUlvE0_clEvENKUlvE18_clEvEUlN3c104HalfEE_NS_6detail5ArrayIPcLi2EEE23TrivialOffsetCalculatorILi1EjESE_NS0_6memory12LoadWithCastILi1EEENSF_13StoreWithCastEEEviT_T0_T1_T2_T3_T4_
                    6.53%  3.77162s     31488  119.78us  4.5760us  354.98us  turing_fp16_s1688gemm_fp16_128x128_ldg8_f2f_stages_32x1_nt
                    6.25%  3.61247s     78410  46.071us  1.6640us  2.8498ms  adam
                    6.21%  3.58839s    186237  19.267us     800ns  1.1974ms  void at::native::vectorized_elementwise_kernel<int=4, at::native::AddFunctor<float>, at::detail::Array<char*, int=3>>(int, float, at::native::AddFunctor<float>)
```