@honnibal 2021-09-06
Wow, so great to have this resolved!

I had also noticed that the gradient clipping was surprisingly expensive, but I was never able to figure out how to make it better.